move handleUpdateChartPanAndZoom callback up, because it's needed by toolbox callbacks
