+++ web/gui/dashboard/src/domains/chart/components/abstract-chart.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useMemo } from "react"
+import React, { useCallback } from "react"
 
 import { setGlobalChartUnderlayAction, setGlobalPanAndZoomAction } from "domains/global/actions"
 import { useDispatch } from "react-redux"
@@ -12,7 +12,6 @@ interface Props {
   chartData: ChartData
   chartDetails: ChartDetails
   chartLibrary: ChartLibraryName
-  chartWidth: number
   colors: {
     [key: string]: string
   }
@@ -22,6 +21,8 @@ interface Props {
   legendFormatValue: ((v: number) => number | string) | undefined
   orderedColors: string[]
   hoveredX: number | null
+  onUpdateChartPanAndZoom: (arg: { after: number, before: number, masterID: string }) => void
+
   setHoveredX: (hoveredX: number | null) => void
   setMinMax: (minMax: [number, number]) => void
   unitsCurrent: string
@@ -34,7 +35,6 @@ export const AbstractChart = ({
   chartData,
   chartDetails,
   chartLibrary,
-  chartWidth,
   colors,
   chartUuid,
   dimensionsVisibility,
@@ -42,6 +42,7 @@ export const AbstractChart = ({
   legendFormatValue,
   orderedColors,
   hoveredX,
+  onUpdateChartPanAndZoom,
   setHoveredX,
   setMinMax,
   unitsCurrent,
@@ -49,83 +50,6 @@ export const AbstractChart = ({
   viewBefore,
 }: Props) => {
   const dispatch = useDispatch()
-
-  // old dashboard persists min duration based on first chartWidth, i assume it's a bug
-  // and will update fixedMinDuration when width changes
-  const fixedMinDuration = useMemo(() => (
-    Math.round((chartWidth / 30) * chartDetails.update_every * 1000)
-  ), [chartDetails.update_every, chartWidth])
-
-  const updateChartPanAndZoom = useCallback(({ after, before, callback }) => {
-    if (before < after) {
-      return
-    }
-    let minDuration = fixedMinDuration
-
-    const currentDuraton = Math.round(viewBefore - viewAfter)
-
-    let afterForced = Math.round(after)
-    let beforeForced = Math.round(before)
-    const viewUpdateEvery = chartData.view_update_every * 1000
-
-    // align them to update_every
-    // stretching them further away
-    afterForced -= afterForced % (viewUpdateEvery)
-    beforeForced += viewUpdateEvery - (beforeForced % viewUpdateEvery)
-
-    // the final wanted duration
-    let wantedDuration = beforeForced - afterForced
-
-    // to allow panning, accept just a point below our minimum
-    if ((currentDuraton - viewUpdateEvery) < minDuration) {
-      minDuration = currentDuraton - viewUpdateEvery
-    }
-
-    // we do it, but we adjust to minimum size and return false
-    // when the wanted size is below the current and the minimum
-    // and we zoom
-    let doCallback = true
-    if (wantedDuration < currentDuraton && wantedDuration < minDuration) {
-      minDuration = fixedMinDuration
-
-      const dt = (minDuration - wantedDuration) / 2
-      beforeForced += dt
-      afterForced -= dt
-      wantedDuration = beforeForced - afterForced
-      doCallback = false
-    }
-
-    const tolerance = viewUpdateEvery * 2
-    const movement = Math.abs(beforeForced - viewBefore)
-
-    if (
-      Math.abs(currentDuraton - wantedDuration) <= tolerance && movement <= tolerance && doCallback
-    ) {
-      return
-    }
-
-    // if (this.current.name === 'auto') {
-    //   this.log(logme + 'caller called me with mode: ' + this.current.name);
-    //   this.setMode('pan');
-    // }
-
-
-    // todo support force_update_at in some way
-    // this.current.force_update_at = Date.now() +
-    // window.NETDATA.options.current.pan_and_zoom_delay;
-    // this.current.force_after_ms = after;
-    // this.current.force_before_ms = before;
-    dispatch(setGlobalPanAndZoomAction({
-      after: afterForced,
-      before: beforeForced,
-      masterID: chartUuid,
-    }))
-
-    if (doCallback && typeof callback === "function") {
-      callback()
-    }
-  }, [chartData.view_update_every, chartUuid, dispatch, fixedMinDuration, viewAfter, viewBefore])
-
 
   const setGlobalChartUnderlay = useCallback(({ after, before, masterID }) => {
     dispatch(setGlobalChartUnderlayAction({ after, before, masterID }))
@@ -149,11 +73,11 @@ export const AbstractChart = ({
       legendFormatValue={legendFormatValue}
       orderedColors={orderedColors}
       hoveredX={hoveredX}
+      onUpdateChartPanAndZoom={onUpdateChartPanAndZoom}
       setGlobalChartUnderlay={setGlobalChartUnderlay}
       setHoveredX={setHoveredX}
       setMinMax={setMinMax}
       unitsCurrent={unitsCurrent}
-      updateChartPanAndZoom={updateChartPanAndZoom}
     />
   )
 }
